// food-channel-processor/src/v1/channel-order-food-token/model/info-token-to-comfirm-channel-order.data.model.ts
export class InfoTokenToComfirmChannelOrderDataModel {
    
    restaurant_id: number;
  
    restaurant_brand_id: number;
  
    channel_order_food_id: number;
  
    channel_branch_id: number;
  
    merchant_id: number;
  
    channel_order_food_token_id: number;
  
    access_token: string;
  
    username: string;
  
    password: string;
  
    x_merchant_token: string;
  
    device_id: string;
  
    device_brand: string;
  
    url_get_branch_detail: string;
  
    url_comfirm_order: string;

    url_login: string;
  
    url_update_device: string;
  
    message_error : string;

    user_id: number;

  
  }
  